Split javascript files
Split functions and scenes
