+++ src/index.js
@@ -1,22 +1,29 @@
 import 'phaser';
-
-//import { SimpleScene2 } from './scenes/simple-scene2';
-//import { SimpleScene22 } from './scenes/green';
-//import { WorldScene } from './scenes/green';
-import { BootScene } from './scenes/world';
 import { WorldScene } from './scenes/world';
-import { BattleScene } from './scenes/battle';
-import { UIScene } from './scenes/battle';
+import { BootScene } from './scenes/BootScene';
+import {UIScene} from './scenes/uiscene'
+import {BattleScene} from './scenes/battle'
 
-import { Unit } from './scenes/battle';
-import { Enemy } from './scenes/battle';
-import { PlayerCharacter } from './scenes/battle';
-import { MenuItem } from './scenes/battle';
-import { HeroesMenu } from './scenes/battle';
-import { EnemiesMenu } from './scenes/battle';
-import { gamescene } from './scenes/battle';
+var config = {
+  type: Phaser.AUTO,
+  parent: 'content',
+  width: 320,
+  height: 240,
+  zoom: 2,
+  pixelArt: true,
+  physics: {
+      default: 'arcade',
+      arcade: {
+          gravity: { y: 0 },
+          debug: false // set to true to view zones
+      }
+  },
+  scene: [
+      BootScene,
+      WorldScene,
+      BattleScene,
+      UIScene
+  ]
+};
 
-
-
-gamescene;
-//game;
+var gamescene = new Phaser.Game(config);
