+++ src/scenes/enemy.js
@@ -1,0 +1,9 @@
+import {Unit} from './unit'
+
+export var Enemy = new Phaser.Class({
+  Extends: Unit,
+  initialize:
+  function Enemy(scene, x, y, texture, frame, type, hp, damage) {
+    Unit.call(this, scene, x, y, texture, frame, type, hp, damage);
+  }
+});
